Build in a pure wayland environment
Note that this change only allow the package to build in a pure wayland
environment but doesn't make it functionnal since some X11 -> Wayland
equivalents doesn't exist yet. The changes that remain to be made in
order to get it working in a pure Wayland environment are marked by the
comment "TO DO WAYLAND". Bug report regarding this issue: PTREL-722.

Change-Id: I3e582c1f531ee0fcdb230c3ab366f0e957a647a9
Signed-off-by: Kévin THIERRY <kevin.thierry@open.eurogiciel.org>